Forward Authentication Service (FAS) Integration.
This implements redirection to an external authentication
service in place of splash.html.

It introduces four new options:

1. fasport - if set, enables FAS redirection and the value is
the port used for access.

2. fasremoteip - if set is the ip address of the FAS, defaults
to the NDS gateway address for the case of FAS running locally to NDS.

3. faspath - the path on FAS to the login page. Default "/"

4. fas_secure_enabled - if set to "1", the client token is not
revealed and it is the responsibility of the FAS to request
this from NDS. If set to "0", the client token is passed in
clear text in the query string to the FAS.

Firewall settings are automatically configured for access to the FAS.

Binauth can also be configured at the same time to allow a local
script to be called on authentication.

Signed-off-by: Rob White <rob@blue-wave.net>